Converted Category Priority to float

+++ NewsletterCurator.Data/Category.cs
@@ -13,6 +13,6 @@ namespace NewsletterCurator.Data
         public string Name { get; set; }
         public List<Newsitem> Newsitems { get; set; }
         [Required]
-        public int Priority { get; set; }
+        public float Priority { get; set; }
     }
 }
